backend/src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name venv --file requirements.txt
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
appdirs=1.4.4=pyhd3eb1b0_0
asgiref=3.4.1=pyhd3eb1b0_0
asttokens=2.0.5=pyhd8ed1ab_0
attrs=21.4.0=pyhd3eb1b0_0
automat=20.2.0=py_0
bcrypt=3.2.0=py39he8ac12f_0
beautifulsoup4=4.11.1=pyha770c72_0
brotlipy=0.7.0=py39hb9d737c_1004
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.4.26=h06a4308_0
certifi=2022.6.15=py39h06a4308_0
cffi=1.15.0=py39hd667e15_1
chardet=5.0.0=py39hf3d152e_0
charset-normalizer=2.1.0=pyhd8ed1ab_0
colorama=0.4.5=pyhd8ed1ab_0
conda=4.13.0=py39hf3d152e_1
conda-build=3.21.9=py39hf3d152e_1
conda-package-handling=1.8.1=py39hb9d737c_1
constantly=15.1.0=pyh2b92418_0
cryptography=35.0.0=py39h95dcef6_2
django=3.2.5=pyhd3eb1b0_0
django-cors-headers=3.13.0=pyhd8ed1ab_0
djangorestframework=3.13.1=pyhd8ed1ab_0
ecdsa=0.17.0=pyhd3eb1b0_0
executing=0.8.3=pyhd8ed1ab_0
filelock=3.7.1=pyhd8ed1ab_0
glob2=0.7=py_0
gmp=6.2.1=h295c915_3
hyperlink=21.0.0=pyhd3eb1b0_0
icu=58.2=he6710b0_3
idna=3.3=pyhd8ed1ab_0
incremental=21.3.0=pyhd3eb1b0_0
jinja2=3.1.2=pyhd8ed1ab_1
krb5=1.19.2=hac12032_0
ld_impl_linux-64=2.38=h1181459_1
libarchive=3.4.2=h62408e4_0
libedit=3.1.20210910=h7f8727e_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h516909a_0
liblief=0.11.5=h9c3ff4c_1
libmemcached=1.0.18=hc46c6b1_1
libpq=12.9=h16c4e8d_3
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lxml=4.8.0=py39h1f438cf_0
lz4-c=1.9.3=h9c3ff4c_1
markupsafe=2.1.1=py39hb9d737c_1
mysql-connector-c=6.1.11=h24aacaa_2
mysqlclient=2.0.3=py39h2531618_1
ncurses=6.3=h5eee18b_3
openssl=1.1.1o=h7f8727e_0
paramiko=2.8.1=pyhd3eb1b0_0
patch=2.7.6=h7f98852_1002
patchelf=0.14.5=h58526e2_0
pexpect=4.8.0=pyhd3eb1b0_3
pip=22.1.2=py39h06a4308_0
pkginfo=1.8.3=pyhd8ed1ab_0
psutil=5.9.1=py39hb9d737c_0
psycopg2=2.8.6=py39h3c74f83_1
ptyprocess=0.7.0=pyhd3eb1b0_2
py-lief=0.11.5=py39he80948d_1
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycosat=0.6.3=py39hb9d737c_1010
pycparser=2.21=pyhd8ed1ab_0
pycryptodome=3.12.0=py39hdd9d17f_0
pyhamcrest=2.0.2=pyhd3eb1b0_2
pylibmc=1.6.1=py39hb9d737c_5
pynacl=1.4.0=py39he8ac12f_1
pyopenssl=22.0.0=pyhd8ed1ab_0
pysocks=1.7.1=py39hf3d152e_5
python=3.9.12=h12debd9_1
python-devtools=0.8.0=pyhd8ed1ab_0
python-libarchive-c=4.0=py39hf3d152e_1
python_abi=3.9=2_cp39
pytz=2022.1=py39h06a4308_0
pyyaml=6.0=py39hb9d737c_4
readline=8.1.2=h7f8727e_1
requests=2.28.1=pyhd8ed1ab_0
ripgrep=13.0.0=h2f28480_2
ruamel_yaml=0.15.80=py39h3811e60_1006
service_identity=18.1.0=pyhd3eb1b0_1
setuptools=61.2.0=py39h06a4308_0
six=1.16.0=pyh6c4a22f_0
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.38.5=hc218d9a_0
sqlparse=0.4.1=py_0
tk=8.6.12=h1ccaba5_0
tqdm=4.64.0=pyhd8ed1ab_0
twisted=22.2.0=py39h7f8727e_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.10=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
yaml=0.2.5=h7f98852_2
zlib=1.2.12=h7f8727e_2
zope=1.0=py39h06a4308_1
zope.interface=5.4.0=py39h7f8727e_0
zstd=1.4.9=ha95c52a_0
